move an outdated class to contrib eventsBasedPTRouter
it's the only place the class is still used
